Create site service for reusable functions
